Add transforms for words that don't change sound
Some contractions (using apostrophes) in the source files are done only
to fit the lyrics into the space afforded by the musical notation, and
are not important metrically (e.g. "tho’" sounds the same as "though").
Add these transforms to the map so that the generated lyrics in TXT/
aren't unnecessarily cluttered with meaningless contractions.