Allow independent PDF and MIDI generation
Simplify scores for midi-only cases

Independent PDF and MIDI generation already works where a `\layout`
block exists, because either a `\layout` or a `\midi` block will be
active in each file. For files that do not have a `\layout` (because
they are markup-only) we need to remove the `\score` entirely in the
PDF build case, and this is now accomplished using `\include`.