Mend undesired syllable alignment
Because the bass-only words "for me" were in Lyrics context three, the
third verse of EOG106 had its syllables aligned to the bass line instead
of the soprano, which looked odd. Putting the bass words in their own
lyrics context resolves this. Now the words "for me" appear below the
bass staff, which wasn't the original look, but which avoids the
problems that could occur if we forced them to appear above the bass
staff but below lyrics contexts one through four (thus leaving a large
vertical gap depending on line breaking).