Amended meter in EOG123
Changed EOG123 to 6—8s since it is compatible with EOG007 and there is
no reason to have a single song with `8. 8. 8. 8. 4. 8.` meter.

(This touches on the perennial question of whether the meter should
describe the tune or the lyrics. This project has sought to use the
meter mainly to describe the tune, with some pragmatic concessions.)