Arrange chorus of EOG080 to the words
The tune of EOG080 canonically is sung to "Jesus is Calling" which has a
chorus with differing rhythm in the bass clef from the treble clef. This
sounds nice in the generated audio but is confusing to singers, who see
extra notes that have no syllables to match them. Arrange the tenor and
bass voices to follow the same rhythm as the soprano and alto.