Revert author for EOG001
In 2013, hymnary.org attributed this song to John J. Overholt :

    https://web.archive.org/web/20130516202951/https://hymnary.org/text/come_tis_jesus_gently_calling_ye_with_ca

but as of 2020-08-09, it is marked as Anonymous.

Probably the reference to John J. Overholt was due to compilation. The
Christian Hymnary (compiled by John J. Overholt) does not contain this
hymn, and so it seems somewhat unlikely that the song text was authored
by him.