Change ties to phrasing slurs in EOG048
Their presence accounts for the minority of verses where two words are replaced
by one in a few spots. Since the majority of verses, as well as the refrain,
have two separate words here, we should not generate MIDIs with ties; phrasing
slurs stand in for ties purely for the vocalist's interpretation.